# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Dokumentasi API Sistem Manajemen Parkir.",
        "title": "Parking API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
...
      "/api/v1/operations/transaction-events": {
        "get": {
          "description": "Retrieve all transaction events with pagination support",
          "consumes": [
            "application/json"
          ],
          "produces": [
            "application/json"
          ],
          "tags": [
            "Operations"
          ],
          "summary": "Get all transaction events",
          "parameters": [
            {
              "type": "integer",
              "default": 1,
              "description": "Page number",
              "name": "page",
              "in": "query"
            },
            {
              "type": "integer",
              "default": 10,
              "description": "Items per page",
              "name": "limit",
              "in": "query"
            },
            {
              "type": "string",
              "description": "Search term",
              "name": "search",
              "in": "query"
            }
          ],
          "responses": {
            "200": {
              "description": "OK",
              "schema": {
                "$ref": "#/definitions/responses.PageResponse"
              }
            },
            "400": {
              "description": "Bad Request",
              "schema": {
                "$ref": "#/definitions/responses.BaseResponse"
              }
            }
          }
        },
        "post": {
          "description": "Create a new transaction event record",
          "consumes": [
            "application/json"
          ],
          "produces": [
            "application/json"
          ],
          "tags": [
            "Operations"
          ],
          "summary": "Create a new transaction event",
          "parameters": [
            {
              "description": "Transaction event data",
              "name": "request",
              "in": "body",
              "required": true,
              "schema": {
                "$ref": "#/definitions/requests.CreateTransactionEventRequest"
              }
            }
          ],
          "responses": {
            "201": {
              "description": "Created",
              "schema": {
                "$ref": "#/definitions/responses.BaseResponse"
              }
            },
            "400": {
              "description": "Bad Request",
              "schema": {
                "$ref": "#/definitions/responses.BaseResponse"
              }
            }
          }
        }
      },
      "/api/v1/operations/transaction-events/{id}": {
        "get": {
          "description": "Retrieve a specific transaction event by its ID",
          "consumes": [
            "application/json"
          ],
          "produces": [
            "application/json"
          ],
          "tags": [
            "Operations"
          ],
          "summary": "Get transaction event by ID",
          "parameters": [
            {
              "type": "string",
              "description": "Transaction event ID",
              "name": "id",
              "in": "path",
              "required": true
            }
          ],
          "responses": {
            "200": {
              "description": "OK",
              "schema": {
                "$ref": "#/definitions/responses.BaseResponse"
              }
            },
            "404": {
              "description": "Not Found",
              "schema": {
                "$ref": "#/definitions/responses.BaseResponse"
              }
            }
          }
        },
        "delete": {
          "description": "Delete a specific transaction event by its ID",
          "consumes": [
            "application/json"
          ],
          "produces": [
            "application/json"
          ],
          "tags": [
            "Operations"
          ],
          "summary": "Delete a transaction event",
          "parameters": [
            {
              "type": "string",
              "description": "Transaction event ID",
              "name": "id",
              "in": "path",
              "required": true
            }
          ],
          "responses": {
            "200": {
              "description": "OK",
              "schema": {
                "$ref": "#/definitions/responses.BaseResponse"
              }
            },
            "404": {
              "description": "Not Found",
              "schema": {
                "$ref": "#/definitions/responses.BaseResponse"
              }
            }
          }
        }
      },
...
      "/financials/refunds": {
        "get": {
          "description": "Get list of refunds with pagination",
          "consumes": [
            "application/json"
          ],
          "produces": [
            "application/json"
          ],
          "tags": [
            "Financials"
          ],
          "summary": "Get All Refunds",
          "parameters": [
            {
              "type": "integer",
              "default": 1,
              "description": "Page number",
              "name": "page",
              "in": "query"
            },
            {
              "type": "integer",
              "default": 10,
              "description": "Items per page",
              "name": "limit",
              "in": "query"
            },
            {
              "type": "string",
              "description": "Search by reason",
              "name": "search",
              "in": "query"
            }
          ],
          "responses": {
            "200": {
              "description": "OK",
              "schema": {
                "$ref": "#/definitions/responses.PageResponse"
              }
            },
            "400": {
              "description": "Bad Request",
              "schema": {
                "$ref": "#/definitions/responses.BaseResponse"
              }
            }
          }
        },
        "post": {
          "consumes": [
            "application/json"
          ],
          "produces": [
            "application/json"
          ],
          "tags": [
            "Financials"
          ],
          "summary": "Create New Refund",
          "parameters": [
            {
              "description": "Refund Request",
              "name": "request",
              "in": "body",
              "required": true,
              "schema": {
                "$ref": "#/definitions/requests.CreateRefundRequest"
              }
            }
          ],
          "responses": {
            "201": {
              "description": "Created",
              "schema": {
                "$ref": "#/definitions/responses.BaseResponse"
              }
            },
            "400": {
              "description": "Bad Request",
              "schema": {
                "$ref": "#/definitions/responses.BaseResponse"
              }
            }
          }
        }
      },
      "/financials/refunds/{id}": {
        "get": {
          "consumes": [
            "application/json"
          ],
          "produces": [
            "application/json"
          ],
          "tags": [
            "Financials"
          ],
          "summary": "Get Refund By ID",
          "parameters": [
            {
              "type": "string",
              "description": "Refund Hash ID",
              "name": "id",
              "in": "path",
              "required": true
            }
          ],
          "responses": {
            "200": {
              "description": "OK",
              "schema": {
                "$ref": "#/definitions/responses.BaseResponse"
              }
            },
            "404": {
              "description": "Not Found",
              "schema": {
                "$ref": "#/definitions/responses.BaseResponse"
              }
            }
          }
        },
        "put": {
          "consumes": [
            "application/json"
          ],
          "produces": [
            "application/json"
          ],
          "tags": [
            "Financials"
          ],
          "summary": "Update Refund",
          "parameters": [
            {
              "type": "string",
              "description": "Refund Hash ID",
              "name": "id",
              "in": "path",
              "required": true
            },
            {
              "description": "Update Refund Request",
              "name": "request",
              "in": "body",
              "required": true,
              "schema": {
                "$ref": "#/definitions/requests.UpdateRefundRequest"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "OK",
              "schema": {
                "$ref": "#/definitions/responses.BaseResponse"
              }
            },
            "400": {
              "description": "Bad Request",
              "schema": {
                "$ref": "#/definitions/responses.BaseResponse"
              }
            }
          }
        },
        "delete": {
          "consumes": [
            "application/json"
          ],
          "produces": [
            "application/json"
          ],
          "tags": [
            "Financials"
          ],
          "summary": "Delete Refund",
          "parameters": [
            {
              "type": "string",
              "description": "Refund Hash ID",
              "name": "id",
              "in": "path",
              "required": true
            }
          ],
          "responses": {
            "200": {
              "description": "OK",
              "schema": {
                "$ref": "#/definitions/responses.BaseResponse"
              }
            },
            "404": {
              "description": "Not Found",
              "schema": {
                "$ref": "#/definitions/responses.BaseResponse"
              }
            }
          }
        }
      },
      "/financials/refunds/{id}/approve": {
        "post": {
          "consumes": [
            "application/json"
          ],
          "produces": [
            "application/json"
          ],
          "tags": [
            "Financials"
          ],
          "summary": "Approve Refund",
          "parameters": [
            {
              "type": "string",
              "description": "Refund Hash ID",
              "name": "id",
              "in": "path",
              "required": true
            },
            {
              "description": "Approve Refund Request",
              "name": "request",
              "in": "body",
              "required": true,
              "schema": {
                "$ref": "#/definitions/requests.ApproveRefundRequest"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "OK",
              "schema": {
                "$ref": "#/definitions/responses.BaseResponse"
              }
            },
            "400": {
              "description": "Bad Request",
              "schema": {
                "$ref": "#/definitions/responses.BaseResponse"
              }
            }
          }
        }
      },
      "/financials/refunds/{id}/reject": {
        "post": {
          "consumes": [
            "application/json"
          ],
          "produces": [
            "application/json"
          ],
          "tags": [
            "Financials"
          ],
          "summary": "Reject Refund",
          "parameters": [
            {
              "type": "string",
              "description": "Refund Hash ID",
              "name": "id",
              "in": "path",
              "required": true
            },
            {
              "description": "Reject Refund Request",
              "name": "request",
              "in": "body",
              "required": true,
              "schema": {
                "$ref": "#/definitions/requests.RejectRefundRequest"
              }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/responses.BaseResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/responses.BaseResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
      "requests.ApproveRefundRequest": {
        "type": "object",
        "required": [
          "approved_by"
        ],
        "properties": {
          "approved_by": {
            "type": "string"
          }
        }
      },
...
      "requests.CreateRefundRequest": {
        "type": "object",
        "required": [
          "amount",
          "payment_id",
          "reason",
          "refunded_by",
          "transaction_id"
        ],
        "properties": {
          "amount": {
            "type": "number",
            "minimum": 0
          },
          "payment_id": {
            "type": "string"
          },
          "reason": {
            "type": "string",
            "maxLength": 255
          },
          "refunded_by": {
            "type": "string"
          },
          "transaction_id": {
            "type": "string"
          }
        }
      },
...
      "requests.CreateTransactionEventRequest": {
        "type": "object",
        "required": [
          "event_type",
          "operator_id",
          "transaction_id"
        ],
        "properties": {
          "event_type": {
            "type": "string",
            "enum": [
              "entry",
              "exit",
              "manual_entry",
              "manual_exit",
              "correction",
              "lost_ticket"
            ]
          },
          "notes": {
            "type": "string",
            "maxLength": 5000
          },
          "operator_id": {
            "type": "string"
          },
          "photo_path": {
            "type": "string",
            "maxLength": 500
          },
          "plate_detected": {
            "type": "string",
            "maxLength": 20
          },
          "rfid_detected": {
            "type": "string",
            "maxLength": 50
          },
          "transaction_id": {
            "type": "string"
          }
        }
      },
...
      "requests.RejectRefundRequest": {
        "type": "object",
        "required": [
          "reason"
        ],
        "properties": {
          "reason": {
            "type": "string",
            "maxLength": 255
          }
        }
      },
...
      "requests.UpdateRefundRequest": {
        "type": "object",
        "properties": {
          "amount": {
            "type": "number",
            "minimum": 0
          },
          "approved_by": {
            "type": "string"
          },
          "reason": {
            "type": "string",
            "maxLength": 255
          },
          "status": {
            "type": "string",
            "enum": [
              "requested",
              "pending_approval",
              "approved",
              "rejected",
              "completed"
            ]
          }
        }
      },
...
        "responses.BaseResponse": {
            "type": "object",
            "properties": {
                "data": {},
                "errors": {},
                "message": {
                    "type": "string"
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
        "responses.Meta": {
            "type": "object",
            "properties": {
                "current_page": {
                    "type": "integer"
                },
                "current_url": {
                    "description": "Navigasi URL",
                    "type": "string"
                },
                "filters": {
                    "type": "object",
                    "additionalProperties": true
                },
                "first_page": {
                    "description": "Navigasi Angka",
                    "type": "integer"
                },
                "first_url": {
                    "type": "string"
                },
                "last_page": {
                    "type": "integer"
                },
                "last_url": {
                    "type": "string"
                },
                "limit": {
                    "type": "integer"
                },
                "next_page": {
                    "type": "integer"
                },
                "next_url": {
                    "type": "string"
                },
                "prev_page": {
                    "description": "Pakai pointer biar bisa null kalau gak ada",
                    "type": "integer"
                },
                "prev_url": {
                    "type": "string"
                },
                "search": {
                    "type": "string"
                },
                "sort": {
                    "description": "Search \u0026 Filter",
                    "type": "string"
                },
                "total_data": {
                    "type": "integer"
                },
                "total_page": {
                    "type": "integer"
                }
            }
        },
        "responses.PageResponse": {
            "type": "object",
            "properties": {
                "data": {},
                "message": {
                    "type": "string"
                },
                "meta": {
                    "$ref": "#/definitions/responses.Meta"
                },
                "success": {
                    "type": "boolean"
                }
            }
        }
    }
}